:recycle: refactor: Replace Program with ProgramWithGates
refs #58
