+++ src/react-app/components/ProgramWithGates.spec.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from "@testing-library/react";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 import "@testing-library/jest-dom/vitest";
-import Program from "@components/Program";
+import Program from "@components/ProgramWithGates";
 import userEvent from "@testing-library/user-event";
 
 // ---------------------------------------------------------------------------
