+++ src/react-app/components/ProgramWithGates.tsx
